week 1 tasks
working through the tasks on week 1

+++ Tasks/01 - Variables and Arithmetic/117-typecasting/117-typecasting.cpp
@@ -10,7 +10,7 @@ int main()
 	//Shift left 56 bits and store in a 64bit variable
 	//
 	//We expect the output to be 0b10101100 00000000 00000000 00000000 00000000 00000000 00000000 00000000
-	unsigned long long x = c << 56;
+	unsigned long long x = (unsigned long long)c << 56;
 	displayAndCompare(x);
 
 	//Solution is in the lab notes
@@ -18,6 +18,6 @@ int main()
 	//This also needs fixing
 	int p = 123;
 	int q = 2;
-	double y = p / q;
+	double y = (double)p / (double)q;
 	printf("y = %f\n", y);
 }
